Renames the EmailTemplate to ServiceEmailTemplate, Updates the import reference of EmailTemplate to ServiceEmailTemplate, Makes the copy of ServiceEmailTemplate to be dynamic as well as the preview and subject using the amount prop
